Refactor extract graph method
